Tile Server is now zipping over the rows and their visit information.
We can generate a visit_row for each cbg/day pair.

Updated the joined_location struct to stay as a string, since CBGs can have leading zeroes.

+++ src/components/server/TileServer.hpp
@@ -18,7 +18,7 @@ namespace components::server {
     class TileServer : public hpx::components::component_base<TileServer> {
 
     public:
-        typedef blaze::DynamicMatrix<uint16_t> visit_matrix;
+        typedef blaze::DynamicMatrix<int> visit_matrix;
         typedef blaze::DynamicMatrix<double> distance_matrix;
 
         TileServer();
